refactor(tui): split settings-view into focused rows and memoise activeFields
- Extract section bodies (preset-bar, ai-row, harness-row) + model + editor +
  mutations + section strip; orchestrator drops 685 → 280 LOC, every sibling
  file < 100 LOC.
- Wrap `activeFields` in `useMemo` to silence the react-hooks/exhaustive-deps
  warning that flagged the bare logical-expression default.
- Migrate the warning / check / cross feedback glyphs from inline literals to
  the matching `glyphs.*` theme tokens.
